+++ esp_cam_sensor/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Disabled manual exposure control in the ov5647_mipi_2lane_24Minput_800x1280_raw8_50fps.h.
 - Added Mira220 MIPI driver.
 - add SP0A39 DVP & SPI camera driver.
+- Added detection for all sensors to the `test_apps\detect` test code.
 
 ## 2.0.1